Commit from JBoss Tools

+++ folha-ejb/src/main/java/br/com/folha/xml/parser/CepXMLParser.java
@@ -64,7 +64,7 @@ public class CepXMLParser extends DefaultHandler{
 		    String value = parameters.getProperty(name);  
 		  
 		    // adiciona com um conector (? ou &)  
-		    // o primeiro È ?, depois s„o &  
+		    // o primeiro √© ?, depois s√£o &  
 		    urlString += (++counter == 1 ? "?" : "&")  
 		        + name  
 		        + "="  
@@ -80,7 +80,7 @@ public class CepXMLParser extends DefaultHandler{
 		
 		conexao.connect();
 		
-		//abre a conex„o para input
+		//abre a conex√£o para input
 		BufferedReader br = new BufferedReader(new InputStreamReader(conexao.getInputStream()));
 		
 		newData = new StringBuffer(10000);
@@ -91,15 +91,18 @@ public class CepXMLParser extends DefaultHandler{
 		br.close();
 		
 		arquivo = new File(caminho + File.separator + "cep"+ cep +".xml");
-		FileOutputStream fos = new FileOutputStream(arquivo);
-		PrintWriter out = new PrintWriter(fos, true);
-		
-		//imprime o cÛdigo resultante
-		out.println((new String(newData)).trim());
-		
-		//imprime o n˙mero do resultado
-		//out.println("Resultado: " + conexao.getResponseCode() + "/" + conexao.getResponseMessage() );
-		fos.close();
+		if (!arquivo.exists()) {
+			FileOutputStream fos = new FileOutputStream(arquivo);
+			PrintWriter out = new PrintWriter(fos, true);
+
+			// imprime o c√≥digo resultante
+			out.println((new String(newData)).trim());
+
+			// imprime o n√∫mero do resultado
+			// out.println("Resultado: " + conexao.getResponseCode() + "/" +
+			// conexao.getResponseMessage() );
+			fos.close();
+		}
 		request.parseXML();
 		return endereco;
 	}
@@ -140,8 +143,8 @@ public class CepXMLParser extends DefaultHandler{
 	
 	/** 
 	 * Indica que o parser achou o fim de uma tag/elemento.
-	 * Este evento fornece o nome do elemento, e tambÈm pode
-	 * fornecer as informaÁıes sobre o namespace.
+	 * Este evento fornece o nome do elemento, e tamb√©m pode
+	 * fornecer as informa√ß√µes sobre o namespace.
 	 */
 	public void endElement(String uri, String localName, String tag)
 			throws SAXException {
@@ -149,7 +152,7 @@ public class CepXMLParser extends DefaultHandler{
 		if (tag.equalsIgnoreCase(WEBSERVICE_CEP)){
 			System.out.println();
 		} 
-		//sen„o, seta os atributos do contato
+		//sen√£o, seta os atributos do contato
 		else if (tag.equalsIgnoreCase(RESULTADO)){
 			endereco.setResultado(Integer.parseInt(valorAtual.toString().trim()));
 		} else if (tag.equalsIgnoreCase(RESULTADO_TXT)){
@@ -172,7 +175,7 @@ public class CepXMLParser extends DefaultHandler{
 	}
 	
 	/**
-	 * Indica que o parser achou algum Texto (InformaÁ„o).
+	 * Indica que o parser achou algum Texto (Informa√ß√£o).
 	 */
 	public void characters(char[] ch, int start, int length) {
 		System.out.println(String.copyValueOf(ch,start,length).trim());
